bug 3034: Tag multiple items at once.
Added a "Tag" button in the toolbar area of the results
page.  This activates an inline form that allows the user
to enter a tag that can then be applied to all currently
selected items.

In addition, a "Tag" link has been added to both the List and
Cart pages that provides the same functionality in those contexts.
In these two cases, the a per-item states appears below the
bibliographic information for each item.

Status that applies to the overall operation, such as failing to
provide a tag, will be shown in an alert dialog box.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>