fix for #918
In authorities, you can create frameworks like for biblio.
But in these it's impossible to delete useless fields, because when you click on
delete button, you go back to the default framework.
SO if you confirm your deletion, it delete your field in the default framework
and so in all the framwork (I think...).

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>