Bug fix to OPAC shelf browsing query statement
If the items.location field was NULL, then the current
SQL query would produce no results.  I have turned this
into a conditional block that removes the location condition
in the query if the location is not specified.

In addition, there was a small change to opac-detail.tmpl
that changed Library to Shelves when the shelf browser was
open.  This removes a potential redundant Library Library
display if Library is contained in the starting_homebranch.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>