Bug #2974 Adding category code to BorrowerMandatoryField produced a "The following fields are mandatory: categorycode"
If you made a column that is implemented by a select box mandatory in
BorrowerMandatoryField, Members.js would throw an error on the first
element.  The original intent of the function was probably to catch
the first option for a drop-down list such as "Salutation" where the
text is blank.  Added code to check for an empty string in the first
element.

You can test this by making salutation (title) mandatory.

Also fixed some typos in some of the comments.  Why?  Because I'm a Librarian.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>