bug 3222: messaging preferences for patron categories
Add the ability to set the default circulation messages preferences
for a patron category.  When the EnhancedMessagingPreferences
syspref is ON, the administrator will be able to define
default patron messaging preferences for a patron category.  When
a new patron record is created (either manually or via a patron
import), the new patron's preferences will be copied from the default
for that patron's category.

Signed-off-by: Daniel Sweeney <daniel.sweeney@liblime.com>
Signed-off-by: Galen Charlton <galen.charlton@liblime.com>