Bug 2266: Could not uncollapse a tag if you turn on advancedMARCeditor.
Changed the tagnum <span> into an anchor and added javascript to expand the marc fields if the number is clicked.  This only worked properly when the html tag was an anchor.
The tagnum property in the css file that was changed is only used in addbiblio.tmpl, eventhough additem.tmpl and authorities.tmpl also use this css file.