Bug 2266: Could not uncollapse a tag if you turn on advancedMARCeditor - Revised.
Added javascript and if/else logic to the addbiblio.tmpl and a corresponding anchor tag in CSS.
If advancedMARCeditor is on, the user now clicks on the tag number to expand the marc fields.
If advancedMARCeditor is off, no functionality has changed.  The user clicks on the marc tag's title to expand the marc fields.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>
Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>