Bug 3274: Patch to fix Holiday Exceptions radio button.
Holiday exceptions only make since for repeatable holidays. Javascript showHoliday function was looking for a 'showTitle' id that no longer existed in the showHolidau form.  Added a hidden input field because of the possibility of re-adding the title text input field at a later time.
Added code to determine whether a holiday as an exception or not, so that the "Delete this program" radio button changes to "Delete this exception" because that is the program's true functionality if the holiday is an exception.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>