Bug 2624: Fix for non-persistent check boxes in check-in screen.
Found this bug when playing with AutomaticItemReturn. The "Forgive overdue charges" and "Dropbox mode" would lose their value the item was transfered (if automatic transfer was off), caught a hold, or caught a hold at another location.
This patch does not completely do away with the problem.  If an item is checked-in in simple mode, then one of the forms mentioned above opens, then the user checks one of the selection boxes, the value is not retained.  Javasript would have to be added to transfer the value to the confirmation form when the checkbox is clicked.

This patch also cleans up a few minor XHTML errors.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>
Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>