Use Google Indic Transliteration API to allow users to transliterate english typed words into Indic languages.
This is available only on the masthead search box in the OPAC.

User has to enter a word in english and press space or return to transliterate the word. The destination language can be selected from a drop down list. User can temporarily turn on/off the feature by using "ctrl +g".

A system preference - GoogleIndicTransliteration can be used to turn on this feature.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>