Bug 3228: Fix for Type: label in XSLT displays.
Empty nodes must be generated in the XML for MARC21slim2OPACResults - typeOf008 and MARC21slim2OPACResults - materialTypeCode when the fields in the Marc record do not exist.  This patch fixes the "Type: " label.  More testing needs to be done to see if any other labels appear in the Opac if the data is not defined in the MARC record.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>