Bug 3347: Inconsistencies with public and private list tables in opac-shelves.tmpl.
This patch fixes inconsistencies with the public and private list tables in opac-shelves.tmpl.  It adds a "new list" link to the public list view, and it adds a "sort by" column.

This patch also removes the toggle variables from the tables in opac-shelves.tmpl and Page.pm, and uses the template __odd__ variable for table highlighting if javascript is turned off.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>