Bug 3317: Adds author and added-author to email sent from basket (Cart).
Adds the author and added author fields to the email that is sent from the shopping cart.  This makes the data sent via email consistent with the data that is viewed in the basket.

There is still a problem with author authority records other than 100 tags appearing correctly in the shopping cart.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>