Bug 3390 Barcodes not displayed in Acro Reader >v7.0
Label pdfs containing barcodes generated by Koha do not display barcodes
properly in Adobe Reader >v7.0.

This does not appear to be a bug in Koha code nor in PDF::Reuse::Barcode code,
but rather in the backward compatibility of Adobe Reader.

This patch changes the barcode embedding method from a Type3
font to graphical embedding which appears to work around the issue.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>