Cannot use the same TMPL_VAR to select ALL options.
Only one select can be selected in this case.  If preselection of search type is
desired, I recommend a clean implementation w/ jquery instead of adding a conditional
for each select.  Please note, that is the only good way to get multiple preselections,
since the same .inc file will be used in each case.  So a single variable like
    <TMPL_IF NAME="index_bc_selected">
cannot be used, since it would set for ALL (3) advanced search dropdowns.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>
Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>