bug 3465: enhance DB update reporting by web installer
This enhances the web installer so that messages from updatedatabase.pl will
be displayed to the user running a database update.  Messages printed to STDOUT
will be displayed as the update report, while messages printed to STDERR will
be displayed as update errors.

This patch introduces  a new module dependency, IPC::Cmd.
IPC::Cmd lets one run an external program such as updatedatabase.pl and easily
capture STDERR and STDOUT for further munging.  IPC::Cmd is core in Perl 5.10
and stable in Perl 5.8.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>