bugfix opac-user stomps mylibraryfirst value in quick search label
the system preference mylibraryfirst is exported by Auth containing the logged
in users branchcode.  It is also exported as the value of the syspref by
opac-user.pl.  The effect is that when mylibraryfirst is on and the user
logged in the text label for the OPAC quick search instead of displaying
'Search (in [branchcode] only)' shows 'Search (in 1 only)'.  Other user pages
are uneffected.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>