(Bug 3527) Removed unnecessary WHERE from UPDATE statement
The WHERE in this statement selected a bunch of different accounttypes, mostly legacy items.
But for sub writeoff to be called in the first place, it must have been an account with a
positive amountoustanding already, and we have the borrowernumber and accountnumber, so this
WHERE was un-needed, and prevented writeoff of locally-defined manual invoice types.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>