bug 2859: add index on bibioitems.issn (DB rev 006)
Because of a query on biblioitems.issn in
C4::Breeding::ImportBreeding(), libraries doing a lot
of bib imports were experiencing performance problems,
which adding an index will alleviate.

A possible longer term solution is to index all
reservoir/import batch records in a Zebra database.

[LL bug 232]

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>

Bumping version number for 3.0.4

Conflicts:

	kohaversion.pl