Bug 3173 Place Hold link visible in OPAC even if syspref turned off
If the syspref RequestOnOpac is turned off, the link to place a hold does not
show in the detail page.  However, the link is still present in the results
list as well as in the basket and in lists.  Modified templates to check
for the syspref and to obey it.

NOTE:  This does not turn off the display of the Holds tab in the OPAC
user account.  I think there are situations where library staff may be
allowed to place holds but users are not allowed to do so through the
OPAC.  Leaving the Holds tab visible shows users what they have on hold,
but does not allow them to place holds.  It does allow users to cancel
holds.  If there is a need to hide this tab, that would probably be
best accomplished by a separate syspref.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>
Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>