MT2116:   Addons to the CSV export
Adds the ability to customize CSV exports through the use of a YAML file.

The following customizations are available :
  - Preprocessing
  - Postprocessing
  - Field-by-field processing

The YAML field should be stored in the tools/csv-profiles/ directory and
named after the id of the CSV profile you want to customize.
An example file is provided in that directory.
(cherry picked from commit 76655b5b941f6f6094f705a296f7b68e64d148df)

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>