Change debarring system (MT2263)
Now when a user is debarred, you set a end date, and can put a comment.
This patch is the first, it need some improvement, but works globaly well.
If an overdue generate a debar, the debar date is set to 9999-12-31.

show debar date on member detail
This patch show debar date, if the member is debarred, and he have a valid date(not unlimited debar).

add debar date in circulation
This patch add the debar date if it's not unlimited, in circulation.pl

rework on member edition
This change the way to edit user debar date/comment, to follow nicolas' mockup.

(bug #MT2263) fix functions for debarring

this fix ismemberblock and patronflags for new debarring system

Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>