Fix for Bug 4262, changing matching rule add big message box
Fix includes a change to the format of the page to consolidate
matching rule settings with matching rule controls. This
reduces potential confusion from the placement of the message
box. Assuming that matching rule settings and controls were
displayed separately for clarity and control, I've added
JavaScript which allows the user to change their mind after
changing a selection (but before submitting). It's a new
interaction which I hope is useful.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>