MT3617 : opac search advanced Library set to my library
Without any consideration of the System preference SearchMyLibraryFirst,
When patron is logged in, the library set is set to the patron's library.
This is not the expected behaviour and is implied by GetBranchesLoop
and the fact that users library is selected by default.

Overwriting this behaviour in opac-search.pl when user not logged in
or SearchMyLibraryFirst not set i.e. mybranch not set