Bug 5173 - allow the discount to be editing on acq order
This allows the discount to be edited on the ACQ order page. If the
discount is edited, the other fields will update as appropriate. Note
that if you re-visit the order later, the discount will be set to 0.00%.
Saving it requires a database change, and so will be included in another
commit.