MT5665: branch_transfer_limits emptyed
A library has reported that branch_transfer_limits already were lost sometimes, but not always

Investigating the problem, we've found that it's the pager that is responsible of this.
If you display only 10 lines (by default), all lines that are hidden won't be sent if you <save>

That's why what you previously entered has been lost!

This patch removes the pager, it's useless anyway: when you want to select the limits, it's convenient to have everything on the page, so no pager needed