(MT #7184) Some bugs in readingrecord
Due to the add of tablesorter and tablesorter.pager the links "Showing All Items" and "Show Last 50 Items Only" are useless.
Some lines in the .pl are useless too.
The sort in the Title column needed a parser due to the <a> tag
I have increased the ddl "entries/page" to allow to see full table.