MT5316: Add possibility to cancel an order receipt
If a received order has a parent order, cancelling receipt is done by adding
the order's quantity to the parent order's quantity and removing the order
line.
If it has no parent, cancelling is done by setting datereceived and
booksellerinvoicenumber to NULL, and quantityreceived to 0.
If AcqCreateItem is set to 'receiving', items linked to this order are
removed.