(MT #6784) Opac logo
Solve a bug regarding the use of LibraryName and opacsmallimage.
Now in each case the page is well printed.
Note : if you don't want to have "Koha Online Catalog" under the logo you can edit a space in LibraryName.
Also be careful LibraryName is used in the pdf for printing basketgroup.