Bug 6273: Add support for recording SIP2 Fee Paid Payment Type
Adds support for recording in Koha accountlines a different accounttype
depending on how the fee was paid at the SIP2 station (cash, credit, etc)

Adds a new param to recordpayment(), $sip_paytype, which is appended to the
'Pay' type if present.  The payment descriptin is also appended with " (via SIP2)"
if this is present.

In order for other scripts to keep working as expected, "eq 'Pay'" needed to
be replaced with a regex comparison "=~ /^Pay/", so that 'Pay' and 'Pay-##' would continue
to group together.

To test:
1.  Make a payment over a SIP2 connection
2.  Check the patron record in the staff client; you should see the modified description
3.  Attempt to print a invoice or a reciept for the borrower; the payment should show up
    where expected