Bug 7362 - allow checkout by item number
This adds a system preference (CircFallbackItemnumber) that allows the
item number to be used instead of a barcode when issuing items.

It will check the barcode first, but if it isn't found, it'll attempt to
use the item number.

There is also a bit of related refactoring going on, removing some of
the assumption from Koha that every item has a barcode, and using that
as a primary key in the circulatory system.

Conflicts:
	C4/Circulation.pm
	C4/SIP/ILS/Transaction/Checkout.pm
	C4/SIP/ILS/Transaction/Renew.pm
	circ/circulation.pl
	installer/data/mysql/sysprefs.sql
	koha-tmpl/intranet-tmpl/prog/en/modules/admin/preferences/circulation.pref