Bug 4222 - allow nonpublicnote to be a mapped DB column
This means that it's possible (and easy) to put non-public notes
(usually 952$x) on a template, same as public notes. It creates a
mapping between the 952$x field and the new database column.

Note that when this is applied, it will only work on newly saved
records.

It is possible this will need to be altered to allow for UNIMARC etc. if
they use 952 differently.